+++ providers/install.rb
@@ -12,10 +12,10 @@ require 'chef/mixin/language'
 include Chef::Mixin::ShellOut
 
 def load_current_resource
-  kb_args = kibana_resources
+  @kibana_resource = new_resource.clone
 end
 
-action :delete do
+action :remove do
   kb_args = kibana_resources
 
   idr = directory kb_args[:install_dir] do
@@ -48,7 +48,7 @@ action :create do
     new_resource.updated_by_last_action(gt.updated_by_last_action?)
 
     ln = link "#{kb_args[:install_dir]}/current" do
-      to "#{kb_args[:install_dir]}/#{kb_args[:git_branch]}"
+      to "#{kb_args[:install_dir]}/#{kb_args[:git_branch]}/src"
     end
     new_resource.updated_by_last_action(ln.updated_by_last_action?)
     node.set['kibana'][kb_args[:name]]['web_dir'] = "#{kb_args[:install_dir]}/current/src"
@@ -58,7 +58,7 @@ action :create do
     case kb_args[:file_type]
     when 'tgz'
       rf = remote_file "#{Chef::Config[:file_cache_path]}/kibana_#{kb_args[:name]}.tar.gz" do
-        # checksum kb_args[:file_checksum]
+        checksum kb_args[:file_checksum]
         source kb_args[:file_url]
         action [:create_if_missing]
       end