+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of kibana.
 
+## 1.3.0:
+* big rewrite to act as library cookbook.
+* not sure if it will be backwards compat yet.
+
 ## 1.2.1:
 * proxy backend healthcheck
 * update proxy templates to match kibana master